:toolbox: Edit project
